+++ Tests/TuistPluginTests/PluginServiceTests.swift
@@ -165,6 +165,67 @@ final class PluginServiceTests: TuistTestCase {
         XCTAssertEqual(plugins, expectedPlugins)
     }
 
+    func test_loadPlugins_WHEN_localTasks() throws {
+        // Given
+        let pluginPath = try temporaryPath()
+        let pluginName = "TestPlugin"
+        let tasksPath = pluginPath.appending(components: Constants.tasksDirectoryName)
+
+        try makeDirectories(tasksPath)
+
+        // When
+        manifestLoader.loadConfigStub = { _ in
+            .test(plugins: [.local(path: .relativeToRoot(pluginPath.pathString))])
+        }
+
+        manifestLoader.loadPluginStub = { _ in
+            ProjectDescription.Plugin(name: pluginName)
+        }
+
+        let config = mockConfig(plugins: [TuistGraph.PluginLocation.local(path: pluginPath.pathString)])
+
+        // Then
+        let plugins = try subject.loadPlugins(using: config)
+        let expectedPlugins = Plugins.test(
+            tasks: [
+                PluginTasks(name: pluginName, path: tasksPath),
+            ]
+        )
+        XCTAssertEqual(plugins, expectedPlugins)
+    }
+
+    func test_loadPlugins_WHEN_gitTasks() throws {
+        // Given
+        let pluginGitUrl = "https://url/to/repo.git"
+        let pluginGitId = "1.0.0"
+        let pluginFingerprint = "\(pluginGitUrl)-\(pluginGitId)".md5
+        let cachedPluginPath = cacheDirectoriesProvider.pluginCacheDirectory.appending(components: pluginFingerprint)
+        let pluginName = "TestPlugin"
+        let tasksPath = cachedPluginPath.appending(components: Constants.tasksDirectoryName)
+
+        try makeDirectories(tasksPath)
+
+        // When
+        manifestLoader.loadConfigStub = { _ in
+            .test(plugins: [ProjectDescription.PluginLocation.git(url: pluginGitUrl, tag: pluginGitId)])
+        }
+
+        manifestLoader.loadPluginStub = { _ in
+            ProjectDescription.Plugin(name: pluginName)
+        }
+
+        let config = mockConfig(plugins: [TuistGraph.PluginLocation.gitWithTag(url: pluginGitUrl, tag: pluginGitId)])
+
+        // Then
+        let plugins = try subject.loadPlugins(using: config)
+        let expectedPlugins = Plugins.test(
+            tasks: [
+                PluginTasks(name: pluginName, path: tasksPath),
+            ]
+        )
+        XCTAssertEqual(plugins, expectedPlugins)
+    }
+
     func test_loadPlugins_WHEN_localTemplate() throws {
         // Given
         let pluginPath = try temporaryPath()